pyproject.toml
[project]
name = "organized"
version = "0.1.0"
dependencies = [
    "beeai",
    "fastapi",
    "uvicorn",
    "pyyaml",
    "pydantic",
    "google-genai>=1.29.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "ruff>=0.12.5",
]
